Add realtime container execution logs for BatchOperator
