+++ airflow/providers/amazon/aws/hooks/batch_waiters.py
@@ -29,6 +29,7 @@ import json
 import sys
 from copy import deepcopy
 from pathlib import Path
+from typing import Callable
 
 import botocore.client
 import botocore.exceptions
@@ -36,6 +37,7 @@ import botocore.waiter
 
 from airflow.exceptions import AirflowException
 from airflow.providers.amazon.aws.hooks.batch_client import BatchClientHook
+from airflow.providers.amazon.aws.utils.task_log_fetcher import AwsTaskLogFetcher
 
 
 class BatchWaitersHook(BatchClientHook):
@@ -184,7 +186,12 @@ class BatchWaitersHook(BatchClientHook):
         """
         return self.waiter_model.waiter_names
 
-    def wait_for_job(self, job_id: str, delay: int | float | None = None) -> None:
+    def wait_for_job(
+        self,
+        job_id: str,
+        delay: int | float | None = None,
+        get_batch_log_fetcher: Callable[[str], AwsTaskLogFetcher | None] | None = None,
+    ) -> None:
         """
         Wait for Batch job to complete.  This assumes that the ``.waiter_model`` is configured
         using some variation of the ``.default_config`` so that it can generate waiters with the
@@ -193,6 +200,9 @@ class BatchWaitersHook(BatchClientHook):
         :param job_id: a Batch job ID
 
         :param delay:  A delay before polling for job status
+
+        :param get_batch_log_fetcher: A method that returns batch_log_fetcher of
+            type AwsTaskLogFetcher or None.
 
         :raises: AirflowException
 
@@ -216,10 +226,20 @@ class BatchWaitersHook(BatchClientHook):
             waiter.config.max_attempts = sys.maxsize  # timeout is managed by Airflow
             waiter.wait(jobs=[job_id])
 
-            waiter = self.get_waiter("JobComplete")
-            waiter.config.delay = self.add_jitter(waiter.config.delay, width=2, minima=1)
-            waiter.config.max_attempts = sys.maxsize  # timeout is managed by Airflow
-            waiter.wait(jobs=[job_id])
+            batch_log_fetcher = None
+            try:
+                if get_batch_log_fetcher:
+                    batch_log_fetcher = get_batch_log_fetcher(job_id)
+                    if batch_log_fetcher:
+                        batch_log_fetcher.start()
+                waiter = self.get_waiter("JobComplete")
+                waiter.config.delay = self.add_jitter(waiter.config.delay, width=2, minima=1)
+                waiter.config.max_attempts = sys.maxsize  # timeout is managed by Airflow
+                waiter.wait(jobs=[job_id])
+            finally:
+                if batch_log_fetcher:
+                    batch_log_fetcher.stop()
+                    batch_log_fetcher.join()
 
         except (botocore.exceptions.ClientError, botocore.exceptions.WaiterError) as err:
             raise AirflowException(err)
